+++ src/Resources/contao/languages/en/default.php
@@ -82,6 +82,11 @@ $GLOBALS['TL_LANG']['ctv']['spalte']['schnitt'] = 'Avg. rating';
 $GLOBALS['TL_LANG']['ctv']['spalte']['wettkampf'] = 'Match';
 $GLOBALS['TL_LANG']['ctv']['spalte']['mannschaftspunkteVoll'] = 'Match points';
 $GLOBALS['TL_LANG']['ctv']['spalte']['brettpunkteVoll'] = 'Board points';
+$GLOBALS['TL_LANG']['ctv']['spalte']['runden'] = 'Rounds';
+$GLOBALS['TL_LANG']['ctv']['spalte']['osb'] = 'OSB';
+$GLOBALS['TL_LANG']['ctv']['spalte']['osbVoll'] = 'Olympiad Sonneborn-Berger';
+$GLOBALS['TL_LANG']['ctv']['spalte']['mpsumme'] = 'MP sum';
+$GLOBALS['TL_LANG']['ctv']['spalte']['mpsummeVoll'] = 'Sum of opponents\' match points';
 
 $GLOBALS['TL_LANG']['ctv']['runde'] = 'Round %s';
 $GLOBALS['TL_LANG']['ctv']['spielfrei'] = 'bye';